refactor: extract system dependencies settings into a dedicated component and add architecture rules documentation.
